use pointer receivers for mode handlers
